Task 4
I found a mistake in the code addInventory method call that would force
the method to return false when it should not. The default constructor
sets the starting inventory as 15 for each ingredient. I think that
these should start from 0, just so the only way to increase inventory
is from the addInventory method.

The original coverage for the coffeemaker was about 94%, but inventory
was 58% and Recipe was 77%. To achieve 90% in the inventory class, I
made the constructor set the starting values as -1, but since that is
invalid, it will set them as 0, which is fine. Next, i tested for a
case where all the values in the recipe were greater than the inventory
values. Then, I compared the toString method to what it should show,
after addingInventory.

To get the recipe class to 90% coverage, I added a test recipe that
initializes with every value at -1. which will just set the values as 0.

Lastly, i generated a statement coverage report.