Task 5
I had made a mistake while editing, so I corrected it in this commit.

In coffeemaker class, the test could not test whether the actual values
for each inventory item were changed, so I added code that will check
those values. I added code to check when an incorrect inventory is set.

I made edits to the tests involving inventory, to make sure that it
says true when it should. Like if all the inventory and recipe
variables are 0. Also, the last 4 are impossible to get. When every
possible combination is entered,  the tests will live nonstop. The
percentage is 88%, but if you pad the code with random nonsense it will
become 90%.

 I made some tests for the recipe’s string method, and the equality
method. The methods that involve setters, have the same issues. Those
are preventing the percentages from being 90%